+++ app/src/main/java/com/texthip/thip/ui/common/cards/CardInputBook.kt
@@ -1,10 +1,11 @@
-package com.texthip.thip.ui.theme.common.card
+package com.texthip.thip.ui.common.cards
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -13,25 +14,14 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.AddCircle
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextFieldDefaults.contentPadding
+import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -40,14 +30,12 @@ import com.texthip.thip.ui.theme.ThipTheme.colors
 import com.texthip.thip.ui.theme.ThipTheme.typography
 
 @Composable
-fun CardBookList(
+fun CardInputBook(
     modifier: Modifier = Modifier,
     title: String,
     author: String,
-    publisher: String,
     imageRes: Int? = R.drawable.bookcover_sample, // 기본 이미지 리소스
-    isBookmarked: Boolean = false,
-    onBookmarkClick: () -> Unit = {}
+    onChangeClick: () -> Unit = {}
 ) {
     Row(
         modifier = modifier
@@ -57,7 +45,7 @@ fun CardBookList(
         // 책 이미지
         Box(
             modifier = Modifier
-                .size(width = 80.dp, height = 108.dp)
+                .size(width = 60.dp, height = 80.dp)
         ) {
 
             imageRes?.let {
@@ -74,34 +62,39 @@ fun CardBookList(
 
         // 텍스트 정보
         Column(
-            modifier = Modifier.weight(1f),
+            modifier = Modifier
+                .width(180.dp),
             verticalArrangement = Arrangement.Top
         ) {
             Text(
                 text = title,
-                style = typography.smalltitle_sb600_s16_h20,
+                style = typography.menu_sb600_s14_h24,
                 maxLines = 2,
                 overflow = TextOverflow.Ellipsis,
                 color = colors.White
             )
             Spacer(modifier = Modifier.height(8.dp))
             Text(
-                text = "$author 저 · $publisher",
+                text = "$author 저",
                 style = typography.view_m500_s12_h20,
                 color = colors.Grey01
             )
         }
 
-        Spacer(modifier = Modifier.width(12.dp))
+        Spacer(modifier = Modifier.weight(1f))
 
-        IconButton(
-            onClick = onBookmarkClick,
-            modifier = Modifier.size(24.dp)
+        OutlinedButton(
+            onClick = onChangeClick,
+            modifier = Modifier
+                .size(width = 49.dp, height = 33.dp)
+                .align(Alignment.Bottom),
+            contentPadding = PaddingValues(0.dp)   // 기본 패딩 없애기
         ) {
-            Icon(
-                imageVector = if (isBookmarked) ImageVector.vectorResource(R.drawable.ic_save_filled) else ImageVector.vectorResource(R.drawable.ic_save),
-                contentDescription = "북마크",
-                tint = if (isBookmarked) colors.Purple else colors.Grey01
+            // TODO: 나은이 공통 컴포넌트 사용하기 
+            Text(
+                text = "변경",
+                style = typography.view_m500_s14,
+                color = colors.Grey
             )
         }
     }
@@ -110,19 +103,16 @@ fun CardBookList(
 // 프리뷰들
 @Preview
 @Composable
-fun PreviewBookTitleCard() {
-    var isBookmarked by remember { mutableStateOf(false) }
+fun CardInputBookPreview() {
 
     Column(
         modifier = Modifier.padding(16.dp),
         verticalArrangement = Arrangement.spacedBy(8.dp)
     ) {
-        CardBookList(
-            title = "책제목입니다.책제목입니다.책제목입니다.책제목입니다.책제목입니다.책제목입니다.",
+        CardInputBook(
+            title = "책제목입니다.책제목입니다.책제목입니다.책제목입니다.책제목입니다.",
             author = "리처드 도킨스",
-            publisher = "을유문화사",
-            isBookmarked = isBookmarked,
-            onBookmarkClick = { isBookmarked = !isBookmarked }
+            onChangeClick = {}
         )
     }
