+++ app/src/main/java/com/texthip/thip/ui/common/cards/CardRoomBook.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.theme.common.card
+package com.texthip.thip.ui.common.cards
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -14,8 +14,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.KeyboardArrowRight
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
