+++ app/src/main/java/com/texthip/thip/ui/common/forms/BookPageTextField.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.theme.common.form
+package com.texthip.thip.ui.common.forms
 
 
 import androidx.compose.foundation.clickable
