+++ app/src/main/java/com/texthip/thip/ui/common/forms/FormTextFieldDefault.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.theme.common.form
+package com.texthip.thip.ui.common.forms
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -25,7 +25,7 @@ import com.texthip.thip.ui.theme.ThipTheme.colors
 import com.texthip.thip.ui.theme.ThipTheme.typography
 
 @Composable
-fun BorderedTextField(
+fun BaseInputTextField(
     modifier: Modifier = Modifier,
     hint: String
 ) {
@@ -47,13 +47,12 @@ fun BorderedTextField(
         shape = RoundedCornerShape(12.dp),
         colors = TextFieldDefaults.colors(
             focusedTextColor = colors.White,
-            focusedIndicatorColor = colors.Grey02,
-            unfocusedIndicatorColor = colors.Grey02,
-            focusedContainerColor = colors.Black00,
-            unfocusedContainerColor = colors.Black00,
-            cursorColor = colors.NeonGreen,
+            focusedIndicatorColor = Color.Transparent,
+            unfocusedIndicatorColor = Color.Transparent,
+            focusedContainerColor = colors.Black,
+            unfocusedContainerColor = colors.Black,
+            cursorColor = colors.NeonGreen
         ),
-
         trailingIcon = {
             if (text.isNotEmpty()) {
                 Icon(
@@ -68,19 +67,20 @@ fun BorderedTextField(
                     contentDescription = "Clear text"
                 )
             }
-        }
+        },
+        singleLine = true
     )
 }
 
 @Composable
 @Preview(showBackground = true, backgroundColor = 0xFF000000, widthDp = 360, heightDp = 200)
-fun BorderedTextFieldPreview() {
+fun InputTextFieldPreviewEmpty() {
     Box(
         modifier = Modifier.size(width = 360.dp, height = 200.dp),
         contentAlignment = Alignment.Center
     ) {
-        BorderedTextField(
-            hint = "가이드 텍스트를 입력"
+        BaseInputTextField(
+            hint = "이곳에 텍스트를 입력하세요"
         )
     }
 }
